requirements.txt
pandas
numpy
tweet-preprocessor
nltk
pylint